requirements.txt
Mako==1.2.4
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
requests==2.25.1
